Extract live-object package
