+++ packages/core/echo/echo-schema/src/ast/schema-validator.test.ts
@@ -8,8 +8,6 @@ import { S } from '@dxos/effect';
 import { getDeep } from '@dxos/util';
 
 import { SchemaValidator } from './schema-validator';
-import { create } from '../handler';
-import { TypedObject } from '../object';
 
 describe('schema-validator', () => {
   describe('validateSchema', () => {
@@ -131,52 +129,6 @@ describe('schema-validator', () => {
         valueToAssign: { any: 'value' },
       });
     });
-
-    test('record', () => {
-      const schema = S.mutable(
-        S.Struct({
-          meta: S.optional(S.mutable(S.Any)),
-          // NOTE: S.Record only supports shallow values.
-          // https://www.npmjs.com/package/@effect/schema#mutable-records
-          // meta: S.optional(S.mutable(S.Record({ key: S.String, value: S.Any }))),
-          // meta: S.optional(S.mutable(S.object)),
-        }),
-      );
-
-      {
-        const object = create(schema, {});
-        (object.meta ??= {}).test = 100;
-        expect(object.meta.test).to.eq(100);
-      }
-
-      {
-        const object = create(schema, {});
-        object.meta = { test: { value: 300 } };
-        expect(object.meta.test.value).to.eq(300);
-      }
-
-      {
-        type Test1 = S.Schema.Type<typeof schema>;
-
-        const object: Test1 = {};
-        (object.meta ??= {}).test = 100;
-        expect(object.meta.test).to.eq(100);
-      }
-
-      {
-        class Test2 extends TypedObject({
-          typename: 'dxos.org/type/FunctionTrigger',
-          version: '0.1.0',
-        })({
-          meta: S.optional(S.mutable(S.Record({ key: S.String, value: S.Any }))),
-        }) {}
-
-        const object = create(Test2, {});
-        (object.meta ??= {}).test = 100;
-        expect(object.meta.test).to.eq(100);
-      }
-    });
-
     test('index signatures', () => {
       for (const value of [42, '42']) {
         validateValueToAssign({
