+++ packages/core/echo/echo-schema/src/handler/index.ts
@@ -3,9 +3,5 @@
 //
 
 export * from './expando';
-export * from './logging-handler';
-export * from './object';
 export * from './ref';
-export * from './typed-handler';
-export * from './untyped-handler';
 export * from './utils';
