+++ packages/core/echo/echo-schema/src/index.ts
@@ -15,6 +15,5 @@ export * from './handler';
 export * from './json';
 export * from './mutable';
 export * from './object';
-export * from './proxy';
 export * from './query';
 export * from './types';
