+++ packages/core/echo/echo-schema/src/mutable/types.ts
@@ -5,11 +5,8 @@
 import { S } from '@dxos/effect';
 
 import { getObjectAnnotation, JsonSchemaType } from '../ast';
-import { create } from '../handler';
-import { createJsonSchema } from '../json';
 import { TypedObject } from '../object';
-import { requireTypeReference } from '../proxy';
-import { type ReactiveObject } from '../types';
+import { requireTypeReference } from '../types';
 
 /**
  * Stored representation of a schema.
@@ -20,22 +17,6 @@ export class StoredSchema extends TypedObject({ typename: 'dxos.org/type/Schema'
   version: S.String,
   jsonSchema: JsonSchemaType,
 }) {}
-
-/**
- * Create ECHO object representing schema.
- */
-export const createStoredSchema = ({
-  typename,
-  version,
-  jsonSchema,
-}: Pick<StoredSchema, 'typename' | 'version'> &
-  Partial<Pick<StoredSchema, 'jsonSchema'>>): ReactiveObject<StoredSchema> => {
-  return create(StoredSchema, {
-    typename,
-    version,
-    jsonSchema: jsonSchema ?? createJsonSchema(),
-  });
-};
 
 /**
  * Wrapper around a schema that is stored in the database (from a type definition) but cannot be modified at runtime.
