+++ packages/core/echo/echo-schema/src/object/typed-object.test.ts
@@ -7,8 +7,6 @@ import { describe, expect, test } from 'vitest';
 import { AST, S } from '@dxos/effect';
 
 import { TypedObject } from './typed-object';
-import { create } from '../handler';
-import { getSchema } from '../proxy';
 
 class OrganizationType extends TypedObject({
   typename: 'example.com/type/Organization',
@@ -16,21 +14,6 @@ class OrganizationType extends TypedObject({
 })({
   name: S.String,
 }) {}
-
-class PersonType extends TypedObject<PersonType>({
-  typename: 'example.com/type/Person',
-  version: '0.1.0',
-})(
-  {
-    name: S.String,
-  },
-  {
-    partial: true,
-    record: true,
-  },
-) {}
-
-const TEST_ORG: Omit<OrganizationType, 'id'> = { name: 'Test' };
 
 describe('EchoObject class DSL', () => {
   test('type is a valid schema', async () => {
@@ -41,32 +24,11 @@ describe('EchoObject class DSL', () => {
     expect(OrganizationType.typename).to.eq('example.com/type/Organization');
   });
 
-  test('static isInstance check', async () => {
-    const obj = create(OrganizationType, TEST_ORG);
-    expect(obj instanceof OrganizationType).to.be.true;
-    expect({ id: '12345', ...TEST_ORG } instanceof OrganizationType).to.be.false;
-  });
-
   test('expect constructor to throw', async () => {
     expect(() => new OrganizationType()).to.throw();
   });
 
   test('expect schema', async () => {
     console.log(AST.isTypeLiteral(OrganizationType.ast));
-  });
-
-  test('can get object schema', async () => {
-    const obj = create(OrganizationType, TEST_ORG);
-    expect(getSchema(obj)).to.deep.eq(OrganizationType);
-  });
-
-  describe('class options', () => {
-    test('can assign undefined to partial fields', async () => {
-      const person = create(PersonType, { name: 'John' });
-      person.name = undefined;
-      person.recordField = 'hello';
-      expect(person.name).to.be.undefined;
-      expect(person.recordField).to.eq('hello');
-    });
   });
 });
