+++ packages/core/echo/live-object/.eslintrc.js
@@ -1,0 +1,9 @@
+module.exports = {
+  "extends": [
+    "../../../../.eslintrc.js"
+  ],
+  "parserOptions": {
+    "project": "tsconfig.json",
+    "tsconfigRootDir": __dirname,
+  }
+}
