+++ packages/core/echo/live-object/ROADMAP.md
@@ -1,0 +1,23 @@
+# ECHO-DB
+
+## Stage 1
+
+- [x] EchoObject with weak references.
+- [x] Queries.
+- [x] Selection.
+- [ ] Unify query/selection.
+- [ ] References (e.g., Task => Person).
+- [ ] ECHO pipeline instantiate EchoObject.
+- [ ] Cards demo (multi-peer).
+- [ ] Stringify (tree hierarchy).
+- [ ] Object.keys.
+
+## Stage 2
+
+- [ ] Schema generated types.
+- [ ] CRM/recruiting demo.
+- [ ] Sets/Ordered list.
+- [ ] Type-safe queries.
+- [ ] Materialized links (outside of object/schema)
+- [ ] Immutable objects.
+- [ ] Batch/transactions.
