+++ packages/core/echo/live-object/package.json
@@ -1,0 +1,53 @@
+{
+  "name": "@dxos/live-object",
+  "version": "0.7.1",
+  "description": "Live objects for ECHO.",
+  "homepage": "https://dxos.org",
+  "bugs": "https://github.com/dxos/dxos/issues",
+  "license": "MIT",
+  "author": "info@dxos.org",
+  "sideEffects": true,
+  "exports": {
+    ".": {
+      "types": "./dist/types/src/index.d.ts",
+      "browser": "./dist/lib/browser/index.mjs",
+      "node": "./dist/lib/node-esm/index.mjs"
+    },
+    "./testing": {
+      "types": "./dist/types/src/testing/index.d.ts",
+      "browser": "./dist/lib/browser/testing/index.mjs",
+      "node": "./dist/lib/node-esm/testing/index.mjs"
+    }
+  },
+  "types": "dist/types/src/index.d.ts",
+  "typesVersions": {
+    "*": {
+      "testing": [
+        "dist/types/src/testing/index.d.ts"
+      ]
+    }
+  },
+  "files": [
+    "dist",
+    "src"
+  ],
+  "dependencies": {
+    "@dxos/echo-protocol": "workspace:*",
+    "@dxos/echo-schema": "workspace:*",
+    "@dxos/echo-signals": "workspace:*",
+    "@dxos/effect": "workspace:*",
+    "@dxos/invariant": "workspace:*",
+    "@dxos/keys": "workspace:*",
+    "@dxos/log": "workspace:*",
+    "@dxos/node-std": "workspace:*",
+    "@dxos/util": "workspace:*",
+    "@effect/schema": "^0.75.5",
+    "@preact/signals-core": "^1.6.0",
+    "effect": "^3.9.2",
+    "ulidx": "^2.3.0"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "beast": {}
+}
