+++ packages/core/echo/live-object/src/proxy/getter.ts
@@ -5,9 +5,9 @@
 import { Reference } from '@dxos/echo-protocol';
 import { type S } from '@dxos/effect';
 
+import type { BaseObject } from '@dxos/echo-schema';
+import { SchemaMetaSymbol } from '@dxos/echo-schema';
 import { getProxyHandler, isReactiveObject } from './proxy';
-import { getObjectAnnotation, SchemaMetaSymbol } from '../ast';
-import type { BaseObject } from '../types';
 
 /**
  * Returns the schema for the given object if one is defined.
@@ -18,22 +18,6 @@ export const getSchema = <T extends BaseObject<T>>(obj: T | undefined): S.Schema
   }
 
   return undefined;
-};
-
-export const getTypeReference = (schema: S.Schema<any> | undefined): Reference | undefined => {
-  if (!schema) {
-    return undefined;
-  }
-
-  const annotation = getObjectAnnotation(schema);
-  if (annotation == null) {
-    return undefined;
-  }
-  if (annotation.schemaId) {
-    return new Reference(annotation.schemaId);
-  }
-
-  return Reference.fromLegacyTypename(annotation.typename);
 };
 
 export const isDeleted = <T extends BaseObject<T>>(obj: T): boolean => {
@@ -60,15 +44,4 @@ export const getTypename = <T extends BaseObject<T>>(obj: T): string | undefined
     return (schema as any)[SchemaMetaSymbol].typename;
   }
   return getType(obj)?.objectId;
-};
-export const getTypenameOrThrow = (schema: S.Schema<any>): string => requireTypeReference(schema).objectId;
-
-export const requireTypeReference = (schema: S.Schema<any>): Reference => {
-  const typeReference = getTypeReference(schema);
-  if (typeReference == null) {
-    // TODO(burdon): Catalog user-facing errors (this is too verbose).
-    throw new Error('Schema must be defined via TypedObject.');
-  }
-
-  return typeReference;
 };
