+++ packages/core/echo/live-object/src/proxy/index.d.ts
@@ -1,0 +1,5 @@
+export * from './array';
+export * from './getter';
+export * from './proxy';
+export * from './types';
+//# sourceMappingURL=index.d.ts.map
