+++ packages/core/echo/live-object/src/proxy/index.ts
