+++ packages/core/echo/live-object/src/proxy/proxy.ts
@@ -6,7 +6,7 @@ import { invariant } from '@dxos/invariant';
 
 import { ReactiveArray } from './array';
 import { type ReactiveHandler } from './types';
-import { type BaseObject, type ReactiveObject } from '../types';
+import { type BaseObject, type ReactiveObject } from '@dxos/echo-schema';
 
 // TODO(burdon): Need tighter tests for these.
 // TODO(burdon): Reconcile Proxy and Reactive Object names.
