+++ packages/core/echo/live-object/src/proxy/types.d.ts
@@ -1,0 +1,25 @@
+import { type Reference } from '@dxos/echo-protocol';
+import { type S } from '@dxos/effect';
+import { type BaseObject, type ObjectMeta } from '@dxos/echo-schema';
+/**
+ * Reactive object proxy.
+ */
+export interface ReactiveHandler<T extends BaseObject<T>> extends ProxyHandler<T> {
+    /**
+     * Target to Proxy mapping.
+     */
+    readonly _proxyMap: WeakMap<object, any>;
+    /**
+     * Called when a proxy is created for this target.
+     */
+    init(target: T): void;
+    isDeleted(target: T): boolean;
+    getSchema(target: T): S.Schema<any> | undefined;
+    /**
+     * We always store a type reference together with an object, but schema might not have been
+     * registered or replicated yet.
+     */
+    getTypeReference(target: T): Reference | undefined;
+    getMeta(target: T): ObjectMeta;
+}
+//# sourceMappingURL=types.d.ts.map
