+++ release-please-config.json
@@ -264,6 +264,11 @@
         },
         {
           "type": "json",
+          "path": "packages/core/echo/live-object/package.json",
+          "jsonpath": "$.version"
+        },
+        {
+          "type": "json",
           "path": "packages/core/functions/package.json",
           "jsonpath": "$.version"
         },
